pyproject.toml
[tool.poetry]
name = "vrc-v2t-osc"
version = "0.1.0"
description = ""
authors = ["phillychi3 <phillychi3@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-osc = "^1.9.3"
pyaudio = "^0.2.14"
textual = "^3.1.0"
textual-dev = "^1.7.0"
numpy = "^2.2.4"
openai-whisper = "^20240930"
torch = { version = "^2.1.0", source = "torch" }
torchaudio = "^2.6.0"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu126"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
